basic view flight search

+++ mobile/emiratec/lib/objects/promotion.dart
@@ -4,7 +4,7 @@ class promotion {
   DateTime startDate;
   DateTime endDate;
   String imgPath;
-  Float percentage;
+  double percentage;
 
   promotion({
     required this.startDate,
